refactor: 헬퍼함수 재사용 및 경로 변경에 따른 수정 #19
